+++ CHANGELOG.md
@@ -22,6 +22,8 @@
   - add `utils.position.find_element_in_list` to get all indices of an element in a list
   - add `utils.segmentation.split_list_by_element` to split a list into segments by one element
   - integrate Learn-REx example
+  - fix select model by training loss bugs in `tasks.simple_task.SimpleTask`
+  - add `utils.segmentation.split_data_with_portion` to help split data into multiple portions
 
 ## Previous versions